update tests for fornax_base
