+++ .github/workflows/run-tests.yml
@@ -54,7 +54,7 @@ jobs:
           image=${{ matrix.image }}
           docker run -i --rm -v .:/opt/workspace/repo-code \
             $image \
-            bash -c "pytest -s -v /opt/workspace/repo-code/tests/test_${image}.py -o cache_dir=/tmp/"
+            bash -c "pytest -s -v /opt/workspace/repo-code/tests/test_${image//-/_}.py -o cache_dir=/tmp/"
     
       - name: "Run jupyter-lab: ${{ matrix.image }}"
         run: |
